+++ src/lib.rs
@@ -25,12 +25,27 @@
 //! when compiled with the following features:
 //!
 //! ### `tokio`
-//! If you enable the `tokio` feature you can use [`Promise::spawn_async`] and [`Promise::spawn_blocking`]
+//! If you enable the `tokio` feature you can use [`Promise::spawn_async`], [`Promise::spawn_local`] and [`Promise::spawn_blocking`]
 //! which will spawn tasks in the surrounding tokio runtime.
 //!
 //! ### `web`
-//! If you enable the `web` feature you can use [`Promise::spawn_async`] which will spawn tasks using
+//! If you enable the `web` feature you can use [`Promise::spawn_local`] which will spawn tasks using
 //! [`wasm_bindgen_futures::spawn_local`](https://rustwasm.github.io/wasm-bindgen/api/wasm_bindgen_futures/fn.spawn_local.html).
+//!
+//! ### `smol`
+//! If you enable the `smol` feature you can use [`Promise::spawn_async`] and [`Promise::spawn_local`]
+//! which will spawn tasks using the smol executor. Remember to tick the smol executor with [`tick`] and [`tick_local`].
+//!
+//! ### `smol_tick_poll`
+//! Enabling the `smol_tick_poll` with `smol` calling [`Promise::poll`] will automatically tick the smol executor.
+//! This means you do not have to worry about calling [`tick`] but comes at the cost of loss of finer control over the executor.
+//!
+//! Since calling [`tick_local`] will block the current thread, running multiple local promises at once with `smol_tick_poll` enabled
+//! may also cause stuttering.
+//!
+//! poll-promise will automatically tick the smol executor with this feature disabled for you when using [`Promise::block_until_ready`]
+//! and friends, however.
+//!
 
 // BEGIN - Embark standard lints v6 for Rust 1.55+
 // do not change or add/remove here, but one can add exceptions after this section
@@ -115,4 +130,144 @@
 
 mod promise;
 
-pub use promise::{Promise, Sender};
+pub use promise::{Promise, Sender, TaskType};
+
+#[cfg(feature = "smol")]
+static EXECUTOR: smol::Executor<'static> = smol::Executor::new();
+#[cfg(feature = "smol")]
+thread_local! {
+    static LOCAL_EXECUTOR: smol::LocalExecutor<'static> = smol::LocalExecutor::new();
+}
+
+/// 'Tick' the smol thread executor.
+///
+/// Poll promise will call this for you when using [`Promise::block_until_ready`] and friends.
+/// If so desired [`Promise::poll`] will run this for you with the `smol_tick_poll` feature.
+#[cfg(feature = "smol")]
+pub fn tick() -> bool {
+    crate::EXECUTOR.try_tick()
+}
+
+/// 'Tick' the smol local thread executor.
+///
+/// Poll promise will call this for you when using [`Promise::block_until_ready`] and friends.
+/// If so desired [`Promise::poll`] will run this for you with the `smol_tick_poll` feature.
+#[cfg(feature = "smol")]
+pub fn tick_local() -> bool {
+    crate::LOCAL_EXECUTOR.with(|exec| exec.try_tick())
+}
+
+#[cfg(test)]
+mod test {
+    use crate::Promise;
+
+    #[test]
+    fn it_spawns_threads() {
+        let promise = Promise::spawn_thread("test", || {
+            std::thread::sleep(std::time::Duration::from_secs(1));
+            0
+        });
+
+        assert_eq!(0, promise.block_and_take());
+    }
+
+    #[test]
+    #[cfg(feature = "smol")]
+    fn it_runs_async_threaded() {
+        let promise = Promise::spawn_async(async move { 0 });
+
+        assert_eq!(0, promise.block_and_take());
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    #[cfg(feature = "tokio")]
+    async fn it_runs_async_threaded() {
+        let promise = Promise::spawn_async(async move { 0 });
+
+        assert_eq!(0, promise.block_and_take());
+    }
+
+    #[test]
+    #[cfg(feature = "smol")]
+    fn it_runs_locally() {
+        let promise = Promise::spawn_local(async move { 0 });
+
+        assert_eq!(0, promise.block_and_take());
+    }
+
+    #[test]
+    #[cfg(feature = "smol")]
+    fn it_runs_background() {
+        let promise = Promise::spawn_async(async move {
+            let mut e = 0;
+            for i in -10000..0 {
+                e += i;
+            }
+            e
+        });
+        #[cfg(not(feature = "smol_tick_poll"))]
+        crate::tick();
+
+        std::thread::sleep(std::time::Duration::from_secs(1));
+        assert!(promise.ready().is_some(), "was not finished");
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    #[cfg(feature = "tokio")]
+    async fn it_runs_background() {
+        let promise = Promise::spawn_async(async move {
+            let mut e = 0;
+            for i in -10000..0 {
+                e += i;
+            }
+            e
+        });
+
+        std::thread::sleep(std::time::Duration::from_secs(1));
+        assert!(promise.ready().is_some(), "was not finished");
+    }
+
+    #[test]
+    #[cfg(feature = "smol")]
+    fn it_can_block() {
+        let promise = Promise::spawn_local(async move {
+            std::thread::sleep(std::time::Duration::from_secs(1));
+        });
+        #[cfg(not(feature = "smol_tick_poll"))]
+        crate::tick_local();
+
+        assert!(promise.ready().is_some(), "was not finished");
+    }
+
+    #[test]
+    #[cfg(feature = "smol")]
+    fn it_can_run_async_functions() {
+        let promise = Promise::spawn_async(async move { something_async().await });
+        #[cfg(not(feature = "smol_tick_poll"))]
+        crate::tick();
+
+        assert!(promise.block_and_take(), "example.com is ipv4");
+    }
+    #[tokio::test(flavor = "multi_thread")]
+    #[cfg(feature = "tokio")]
+    async fn it_can_run_async_functions() {
+        let promise = Promise::spawn_async(async move { something_async().await });
+
+        assert!(promise.block_and_take(), "example.com is ipv4");
+    }
+
+    #[test]
+    #[cfg(feature = "smol")]
+    fn it_can_run_async_functions_locally() {
+        let promise = Promise::spawn_local(async move { something_async().await });
+        #[cfg(not(feature = "smol_tick_poll"))]
+        crate::tick_local();
+
+        assert!(promise.block_and_take(), "example.com is ipv4");
+    }
+
+    #[cfg(any(feature = "smol", feature = "tokio"))]
+    async fn something_async() -> bool {
+        async_net::resolve("example.com:80").await.unwrap()[0].is_ipv4()
+    }
+}