+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `smol` feature to enable the use of [`smol`](https://github.com/smol-rs/smol)
+- refactor `Promise::spawn_async` into two new functions, `Promise::spawn_async` and `Promise::spawn_local`
+- `smol_tick_poll` feature to automatically tick the smol executor when polling promises
 
-
-## [0.2.0] - 2022-01-10
+## [0.2.0] - 2022-10-25
 ### Added
 - `web` feature to enable `Promise::spawn_async` using [`wasm_bindgen_futures::spawn_local`](https://rustwasm.github.io/wasm-bindgen/api/wasm_bindgen_futures/fn.spawn_local.html).
 - Add `Promise::abort` to abort the associated async task, if any.
@@ -18,5 +21,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial commit - add the `Promise` type.
 
 [Unreleased]: https://github.com/EmbarkStudios/poll-promise/compare/0.2.0...HEAD
-[0.1.0]: https://github.com/EmbarkStudios/poll-promise/releases/tag/0.2.0
+[0.2.0]: https://github.com/EmbarkStudios/poll-promise/releases/tag/0.2.0
 [0.1.0]: https://github.com/EmbarkStudios/poll-promise/releases/tag/0.1.0